SetQuery支持多参数
