+++ _example/go.mod
@@ -3,6 +3,7 @@ module main
 go 1.13
 
 require (
-	github.com/gin-gonic/gin v1.6.2 // indirect
-	github.com/guonaihong/gout v0.0.11 // indirect
+	github.com/gin-gonic/gin v1.6.2
+	github.com/google/uuid v1.1.1
+	github.com/guonaihong/gout v0.0.11
 )
